Update LICENSE to MIT
As requested I've decided to change the license of the project to the more permissive and open MIT license. As far as I'm aware the MIT license is compatible with the GPL license but if anyone has any issues, feel free to open and issue or contact me. This closes issue #8.